Changed the Activity for the CheatFunction
