+++ app/src/main/java/com/group7/dominion/GameActivity.java
@@ -33,8 +33,7 @@ import java.util.ArrayList;
 
 
 public class GameActivity extends AppCompatActivity {
-    private SensorManager sm;
-    private ShakeListener shakeListener;
+
     private TabLayout tabLayout;
     private ViewPager viewPager;
     private Toolbar toolbar;
@@ -60,69 +59,17 @@ public class GameActivity extends AppCompatActivity {
 
 
 
-        sendUpdateMessage();
-        ArrayList<String> names = new ArrayList<>();
-        ClientConnector.getClientConnector().registerCallback(UpdatePlayerNamesMsg.class, (msg -> {
-            runOnUiThread(() -> {
-                names.clear();
-                names.addAll(((UpdatePlayerNamesMsg) msg).getNameList());
-            });
-        }));
-        shakeListener = new ShakeListener(getSupportFragmentManager(), getUsername(), names);
-        sm = (SensorManager) getSystemService(SENSOR_SERVICE);
-        sm.registerListener(shakeListener.newSensorListener(), sm.getDefaultSensor(Sensor.TYPE_ACCELEROMETER), SensorManager.SENSOR_DELAY_NORMAL);
 
 
     }
 
 
-    public void sendUpdateMessage() {
-        Thread th = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                ClientConnector.getClientConnector().updatePlayerNames();
-            }
-        });
-        th.start();
-    }
 
-    public String getUsername() {
-        SharedPreferences sharedPreferences = getSharedPreferences("USERNAME", Context.MODE_PRIVATE);
-        String str = sharedPreferences.getString("us", null);
-        return str;
-
-    }
 
     @Override
     protected void onStart() {
         super.onStart();
 
         // Handle communication with Server, only send updated to server whenever card is played etc.
-        ClientConnector clientConnector = ClientConnector.getClientConnector();
-        Game clientGame = clientConnector.getGame();
-        //clientConnector.startGame(); // Send Server Message to start game logic
-        // TODO display playerlist -> Check features
-        // TODO create board and display cards
-
-        clientConnector.registerCallback(HasCheatedMessage.class, (msg -> {
-            runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    String CheaterName = ((HasCheatedMessage) msg).getName();
-                    Toast.makeText(getApplicationContext(), CheaterName + " hat eine zusätzliche Karte gezogen...", Toast.LENGTH_SHORT).show();
-                }
-            });
-        }));
-
-        clientConnector.registerCallback(SuspectMessage.class, (msg -> {
-            runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    String SuspectedUser = ((SuspectMessage) msg).getSuspectedUserName();
-                    String Username = ((SuspectMessage)msg).getUserName();
-                    Toast.makeText(getApplicationContext(), Username + " glaubt, dass " + SuspectedUser + " geschummelt hat", Toast.LENGTH_SHORT).show();
-                }
-            });
-        }));
     }
 }
